data-processing/spark/pyproject.toml
[project]
name = "spark-quick-start"
version = "0.0.1"
description = "Quick Start guide for Spark and derivatives"
requires-python = ">=3.12,<3.13"
readme = "README.md"
authors = [
    {name = "Denis Arnaud", email = "denis.arnaud_fedora@m4x.org"},
]
urls = {repository = "https://github.com/data-engineering-helpers/ks-cheat-sheets"}


# DEPENDENCIES
dependencies = [
    "pyspark",
    "delta-spark",
    "loguru",
    "PyYAML"
]

[dependency-groups]
dev = [
  "packaging",
  "ruff",
  "setuptools",
  "toml",
  "twine",
]
lint = [
  "ruff"
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project.scripts]
#spark-poc = "spark_poc.jobs:cli"

[tool.ruff]
#src = ["src"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "I"]
ignore = ["E203", "E501", "E712"]

[tool.ruff.lint.pydocstyle]
convention = "google"

